pyproject.toml
[tool.poetry]
name = "repl-nix-hbcuexcel"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_hbcuexcel"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
pymysql = "^1.1.1"
werkzeug = "^3.0.3"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
langchain-community = "^0.2.11"
langchain-openai = "^0.1.20"
langchain = "^0.2.12"
openai = "^1.40.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
